Implement AddressController and service

+++ backend/src/main/java/no/ntnu/idatt2106/krisefikser/repository/AddressRepository.java
@@ -8,6 +8,6 @@ import no.ntnu.idatt2106.krisefikser.model.Address;
  * Repository class for managing address data in the database. 
  * This class is responsible for performing CRUD operations on address entities.
  */
-public interface AddressRepository extends JpaRepository<Address, Long> {
+public interface AddressRepository extends JpaRepository<Address, Integer> {
     
 }
